very mnogo changes

+++ Clinic/Windows/AdmWindow.xaml.cs
@@ -117,6 +117,32 @@ namespace Clinic.Windows
                 MessageBox.Show("Выберите сотрудника для редактирования.", "Ошибка", MessageBoxButton.OK, MessageBoxImage.Error);
             }
         }
+
+        private bool IsEmployeeInRecordTable(string login)
+        {
+            using (SqlConnection connection = new SqlConnection(connectionString))
+            {
+                string checkRecordQuery = "SELECT COUNT(*) FROM Record WHERE UserID = (SELECT UserID FROM [User] WHERE Login = @Login)";
+                try
+                {
+                    connection.Open();
+
+                    using (SqlCommand checkRecordCommand = new SqlCommand(checkRecordQuery, connection))
+                    {
+                        checkRecordCommand.Parameters.AddWithValue("@Login", login);
+                        int recordCount = (int)checkRecordCommand.ExecuteScalar();
+
+                        return recordCount > 0;
+                    }
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show("Ошибка при проверке наличия сотрудника в таблице Record: " + ex.Message, "Ошибка", MessageBoxButton.OK, MessageBoxImage.Error);
+                    return false;
+                }
+            }
+        }
+
         private void DeleteEmployeeFromDatabase(string login)
         {
             using (SqlConnection connection = new SqlConnection(connectionString))
@@ -176,10 +202,16 @@ namespace Clinic.Windows
 
             if (result == MessageBoxResult.Yes)
             {
-                employeesList.Remove(selectedEmployee);
+                if (IsEmployeeInRecordTable(selectedEmployee.Login))
+                {
+                    MessageBox.Show("Невозможно удалить.\nУ сотрудника имеется актуальная запись.", "Предупреждение", MessageBoxButton.OK, MessageBoxImage.Warning);
+                    return;
+                }
 
+                employeesList.Remove(selectedEmployee);
                 DeleteEmployeeFromDatabase(selectedEmployee.Login);
             }
+
         }
         private void BackWindow_Click(object sender, RoutedEventArgs e)
         {
@@ -324,6 +356,32 @@ namespace Clinic.Windows
                 MessageBox.Show("Выберите клиента для редактирования", "Ошибка", MessageBoxButton.OK, MessageBoxImage.Error);
             }
         }
+
+        private bool IsClientInRecordTable(int clientID)
+        {
+            using (SqlConnection connection = new SqlConnection(connectionString))
+            {
+                string checkRecordQuery = "SELECT COUNT(*) FROM Record WHERE ClientID = @ClientID";
+                try
+                {
+                    connection.Open();
+
+                    using (SqlCommand checkRecordCommand = new SqlCommand(checkRecordQuery, connection))
+                    {
+                        checkRecordCommand.Parameters.AddWithValue("@ClientID", clientID);
+                        int recordCount = (int)checkRecordCommand.ExecuteScalar();
+
+                        return recordCount > 0;
+                    }
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show("Ошибка при проверке наличия клиента в таблице Record: " + ex.Message, "Ошибка", MessageBoxButton.OK, MessageBoxImage.Error);
+                    return false;
+                }
+            }
+        }
+
         private void DeleteClient_Click(object sender, RoutedEventArgs e)
         {
             if (dgClients.SelectedItem != null)
@@ -336,24 +394,33 @@ namespace Clinic.Windows
                     {
                         Client selectedClient = (Client)dgClients.SelectedItem;
 
-                        using (SqlConnection connection = new SqlConnection(connectionString))
+                        if (selectedClient != null)
                         {
-                            connection.Open();
-
-                            string deleteQuery = "DELETE FROM Client WHERE ClientID = @ClientID";
-                            using (SqlCommand deleteCommand = new SqlCommand(deleteQuery, connection))
+                            if (IsClientInRecordTable(selectedClient.ClientID))
                             {
-                                deleteCommand.Parameters.AddWithValue("@ClientID", selectedClient.ClientID);
-                                int rowsAffected = deleteCommand.ExecuteNonQuery();
+                                MessageBox.Show("Невозможно удалить.\nУ клиента имеется актуальная запись.", "Предупреждение", MessageBoxButton.OK, MessageBoxImage.Warning);
+                                return;
+                            }
 
-                                if (rowsAffected > 0)
+                            using (SqlConnection connection = new SqlConnection(connectionString))
+                            {
+                                connection.Open();
+
+                                string deleteQuery = "DELETE FROM Client WHERE ClientID = @ClientID";
+                                using (SqlCommand deleteCommand = new SqlCommand(deleteQuery, connection))
                                 {
-                                    MessageBox.Show("Клиент успешно удален", "Успех", MessageBoxButton.OK, MessageBoxImage.Information);
-                                    LoadClients();
-                                }
-                                else
-                                {
-                                    MessageBox.Show("Не удалось удалить клиента", "Ошибка", MessageBoxButton.OK, MessageBoxImage.Error);
+                                    deleteCommand.Parameters.AddWithValue("@ClientID", selectedClient.ClientID);
+                                    int rowsAffected = deleteCommand.ExecuteNonQuery();
+
+                                    if (rowsAffected > 0)
+                                    {
+                                        MessageBox.Show("Клиент успешно удален", "Успех", MessageBoxButton.OK, MessageBoxImage.Information);
+                                        LoadClients();
+                                    }
+                                    else
+                                    {
+                                        MessageBox.Show("Не удалось удалить клиента", "Ошибка", MessageBoxButton.OK, MessageBoxImage.Error);
+                                    }
                                 }
                             }
                         }
@@ -369,6 +436,7 @@ namespace Clinic.Windows
                 MessageBox.Show("Выберите клиента для удаления.", "Предупреждение", MessageBoxButton.OK, MessageBoxImage.Warning);
             }
         }
+
         private void SearchClients(string searchString)
         {
             List<Client> filteredClients = new List<Client>();
@@ -477,6 +545,30 @@ namespace Clinic.Windows
         }
 
         // ОКНО ЛЕКАРСТВА
+        private bool IsDrugUsedInConclusion(int drugID)
+        {
+            using (SqlConnection connection = new SqlConnection(connectionString))
+            {
+                string checkConclusionQuery = "SELECT COUNT(*) FROM Conclusion WHERE DrugsID = @DrugsID";
+                try
+                {
+                    connection.Open();
+
+                    using (SqlCommand checkConclusionCommand = new SqlCommand(checkConclusionQuery, connection))
+                    {
+                        checkConclusionCommand.Parameters.AddWithValue("@DrugsID", drugID);
+                        int conclusionCount = (int)checkConclusionCommand.ExecuteScalar();
+
+                        return conclusionCount > 0;
+                    }
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show("Ошибка при проверке использования лекарства в заключениях: " + ex.Message, "Ошибка", MessageBoxButton.OK, MessageBoxImage.Error);
+                    return false;
+                }
+            }
+        }
 
         public void LoadDrugs()
         {
@@ -535,9 +627,9 @@ namespace Clinic.Windows
                     {
                         Drug selectedDrug = (Drug)dgDrugs.SelectedItem;
 
-                        if (IsDrugUsedInDiagnosis(selectedDrug.DrugsID))
+                        if (IsDrugUsedInConclusion(selectedDrug.DrugsID))
                         {
-                            MessageBox.Show("Лекарство нельзя удалить, так как есть диагноз, который использует это лекарство", "Предупреждение", MessageBoxButton.OK, MessageBoxImage.Warning);
+                            MessageBox.Show("Лекарство нельзя удалить, так как есть заключение, которое использует это лекарство", "Предупреждение", MessageBoxButton.OK, MessageBoxImage.Warning);
                             return;
                         }
 
@@ -545,10 +637,10 @@ namespace Clinic.Windows
                         {
                             connection.Open();
 
-                            string deleteQuery = "DELETE FROM Drugs WHERE DrugsID = @DrugID";
+                            string deleteQuery = "DELETE FROM Drugs WHERE DrugsID = @DrugsID";
                             using (SqlCommand deleteCommand = new SqlCommand(deleteQuery, connection))
                             {
-                                deleteCommand.Parameters.AddWithValue("@DrugID", selectedDrug.DrugsID);
+                                deleteCommand.Parameters.AddWithValue("@DrugsID", selectedDrug.DrugsID);
                                 int rowsAffected = deleteCommand.ExecuteNonQuery();
 
                                 if (rowsAffected > 0)
@@ -571,25 +663,10 @@ namespace Clinic.Windows
             }
             else
             {
-                MessageBox.Show("Выбери лекарство для удаления.", "Предупреждение", MessageBoxButton.OK, MessageBoxImage.Warning);
+                MessageBox.Show("Выберите лекарство для удаления.", "Предупреждение", MessageBoxButton.OK, MessageBoxImage.Warning);
             }
         }
-        private bool IsDrugUsedInDiagnosis(int drugID)
-        {
-            using (SqlConnection connection = new SqlConnection(connectionString))
-            {
-                connection.Open();
 
-                string query = "SELECT COUNT(*) FROM Diagnosis WHERE DrugsID = @DrugID";
-                using (SqlCommand command = new SqlCommand(query, connection))
-                {
-                    command.Parameters.AddWithValue("@DrugID", drugID);
-                    int count = (int)command.ExecuteScalar();
-
-                    return count > 0;
-                }
-            }
-        }
         private void SearchDrugs(string searchString)
         {
             List<Drug> filteredDrugs = new List<Drug>();
@@ -647,6 +724,31 @@ namespace Clinic.Windows
 
 
         // ОКНО ДИАГНОЗ
+
+        private bool IsDiagnosisUsedInConclusion(int diagnosisID)
+        {
+            using (SqlConnection connection = new SqlConnection(connectionString))
+            {
+                string checkConclusionQuery = "SELECT COUNT(*) FROM Conclusion WHERE DiagnosisID = @DiagnosisID";
+                try
+                {
+                    connection.Open();
+
+                    using (SqlCommand checkConclusionCommand = new SqlCommand(checkConclusionQuery, connection))
+                    {
+                        checkConclusionCommand.Parameters.AddWithValue("@DiagnosisID", diagnosisID);
+                        int conclusionCount = (int)checkConclusionCommand.ExecuteScalar();
+
+                        return conclusionCount > 0;
+                    }
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show("Ошибка при проверке использования диагноза в заключениях: " + ex.Message, "Ошибка", MessageBoxButton.OK, MessageBoxImage.Error);
+                    return false;
+                }
+            }
+        }
 
         public void LoadDiagnoses()
         {
@@ -760,6 +862,12 @@ namespace Clinic.Windows
                     {
                         Diagnosis selectedDiagnosis = (Diagnosis)dgDiagnosis.SelectedItem;
 
+                        if (IsDiagnosisUsedInConclusion(selectedDiagnosis.DiagnosisID))
+                        {
+                            MessageBox.Show("Диагноз нельзя удалить, так как есть заключение, которое использует этот диагноз", "Предупреждение", MessageBoxButton.OK, MessageBoxImage.Warning);
+                            return;
+                        }
+
                         using (SqlConnection connection = new SqlConnection(connectionString))
                         {
                             connection.Open();
@@ -793,6 +901,7 @@ namespace Clinic.Windows
                 MessageBox.Show("Выберите диагноз для удаления.", "Предупреждение", MessageBoxButton.OK, MessageBoxImage.Warning);
             }
         }
+
 
         // ОКНО ЗАПИСЬ
 
@@ -838,7 +947,30 @@ namespace Clinic.Windows
 
             dgRecords.ItemsSource = records;
         }
+        private bool IsRecordUsedInConclusion(int recordID)
+        {
+            using (SqlConnection connection = new SqlConnection(connectionString))
+            {
+                string checkConclusionQuery = "SELECT COUNT(*) FROM Conclusion WHERE RecordID = @RecordID";
+                try
+                {
+                    connection.Open();
 
+                    using (SqlCommand checkConclusionCommand = new SqlCommand(checkConclusionQuery, connection))
+                    {
+                        checkConclusionCommand.Parameters.AddWithValue("@RecordID", recordID);
+                        int conclusionCount = (int)checkConclusionCommand.ExecuteScalar();
+
+                        return conclusionCount > 0;
+                    }
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show("Ошибка при проверке использования записи в заключениях: " + ex.Message, "Ошибка", MessageBoxButton.OK, MessageBoxImage.Error);
+                    return false;
+                }
+            }
+        }
         private void ContextMenu_AddConclusion_Click(object sender, RoutedEventArgs e)
         {
             if (dgRecords.SelectedItem != null)
@@ -891,6 +1023,12 @@ namespace Clinic.Windows
                     {
                         Record selectedRecord = (Record)dgRecords.SelectedItem;
 
+                        if (IsRecordUsedInConclusion(selectedRecord.RecordID))
+                        {
+                            MessageBox.Show("Запись нельзя удалить, так как есть заключение, которое использует эту запись", "Предупреждение", MessageBoxButton.OK, MessageBoxImage.Warning);
+                            return;
+                        }
+
                         using (SqlConnection connection = new SqlConnection(connectionString))
                         {
                             connection.Open();
@@ -924,6 +1062,7 @@ namespace Clinic.Windows
                 MessageBox.Show("Выберите запись для удаления.", "Предупреждение", MessageBoxButton.OK, MessageBoxImage.Warning);
             }
         }
+
         private void SearchRecords(string searchString)
         {
             List<Record> filteredRecords = new List<Record>();
@@ -1022,7 +1161,5 @@ namespace Clinic.Windows
 
             dgList.ItemsSource = conclusions;
         }
-
-
     }
 }
